taxinnovation/apps/endpoint_api_listo/requirements.txt
Flask==1.1.2
gunicorn==20.0.4 
aiohttp==3.7.2
asyncio==3.4.3
psycopg2-binary==2.8.6
requests==2.25.0 
janus==0.6.1
google-cloud-logging==1.15.1
pandas
